Fixing the Mimer behavior for identifying whether the content should be processed and placed in request.data or kept in the request.POST/PUT.
Most systems do send  a content-type for form-encoded data and that is "application/x-www-form-urlencoded". This resulted in request.content_type to always have a value and not be usable in the handlers.

The request.content_type is now always set to None for normal form-encoded POST submissions, to match the documentation